requirements.txt
librosa
numpy
scipy
torch
pandas
PyYAML
